+++ packages/terra-popup/src/terra-dev-site/doc/example/PopupInsideModal.jsx
@@ -2,10 +2,12 @@ import React from 'react';
 import ModalManager from 'terra-modal-manager';
 import ModalContainer from 'terra-popup/lib/terra-dev-site/doc/example/PopupModalContainer';
 
-const ModalManagerDemo = () => (
-  <ModalManager>
-    <ModalContainer />
-  </ModalManager>
-);
+function ModalManagerDemo() {
+  return (
+    <ModalManager>
+      <ModalContainer />
+    </ModalManager>
+  );
+}
 
 export default ModalManagerDemo;