+++ packages/terra-embedded-content-consumer/CHANGELOG.md
@@ -4,6 +4,11 @@ ChangeLog
 Unreleased
 ----------
 
+3.19.0 - (December 10, 2019)
+------------------
+### Added
+* Added reference wdio screenshots for "orion-fusion-theme".
+
 3.18.0 - (November 7, 2019)
 ------------------
 ### Changed